% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timelines.R
\docType{data}
\name{default_columns_mapping}
\alias{default_columns_mapping}
\title{Default column mapping used for timelines module.}
\format{
An object of class \code{list} of length 2.
}
\usage{
default_columns_mapping
}
\description{
Default column mapping used for timelines module.
}
\keyword{datasets}
