% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timelines.R
\name{shiny_timelines_server}
\alias{shiny_timelines_server}
\title{Timeline component Server function}
\usage{
shiny_timelines_server(
  input,
  output,
  session,
  demographics = NULL,
  lineoftherapy = NULL,
  indications = NULL,
  visits = NULL,
  columns_mapping = default_columns_mapping,
  extra_data = NULL,
  .list = list()
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{demographics}{Unevaluated reactive (\code{shiny::reactive}, \code{shiny::reactiveVal})
value containing demographics data.frame.}

\item{lineoftherapy}{Unevaluated reactive (\code{shiny::reactive}, \code{shiny::reactiveVal})
value containing lineoftherapy data.frame.}

\item{indications}{Unevaluated reactive (\code{shiny::reactive}, \code{shiny::reactiveVal})
value containing indications data.frame.}

\item{visits}{Unevaluated reactive (\code{shiny::reactive}, \code{shiny::reactiveVal})
value containing visits data.frame.}

\item{columns_mapping}{Columns mapping used for the module.}

\item{extra_data}{Named character vector (label - value pairs) of extra columns stored in demographics
to be presented in timeline.}

\item{.list}{(optional) List storing \code{demographics}, \code{lineoftherapy}, \code{visits}, \code{indications} and \code{column_mapping}.
Makes calling module easier in programmatic use (e.g. storing data as \code{reactiveValues}).}
}
\description{
Timeline component Server function
}
