% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimelines.R
\docType{package}
\name{shinyTimelines-package}
\alias{shinyTimelines}
\alias{shinyTimelines-package}
\title{shinyTimelines: Timelines component for dashboards}
\description{
sihnyTimelines module
}
\author{
\strong{Maintainer}: Kamil Wais \email{kamil.wais@gmail.com}

Other contributors:
\itemize{
  \item Krystian Igras \email{krystian8207@gmail.com} [contributor]
}

}
